Checkbox Functionality Update
-After asking for help, I get it now. Progress: getting there.
Key Takeaway: The problem why the method i was using before was not working was bc i was using one checkbox for multiple  tasks that i was adding and thus it made it difficult to track the completion status of each task. So then the solution was; dynamically add a checkbox for each item that gets added as a task. That way it's easier to track its completion status bc its checkbox would be next to it and wouldn't affect the logic for the next tasks that re going to be added .