Filtering Functionality.
-After seeking help, I get it.
Key Takeaway:
The issue i was having was not knowing how to create a connection between the different filtering options and the tasks that i was adding. The solution was creating a switch case method that adds task that have been entered into a list/array that holds them and display their status based on a toggle. If the task are completed, it show the relevant task as checked off and with UI feedback.