Linear Search Input Field Functionality Update Part 3
-Key issue: the searched  number was returning all the numbers in the array instead of the only one searched one.
Key takeaway: instead of using a switch case; after seeking help , i used a for loop instead. It compares the value that is searched with the numbers in the array via their index. for each value that matches the value in the array by index, it returns only that result.
-I also figured out also how to show a message that displays "Value not found" if a value does not exist in the array by first comparing the searched value against numbers in the array by index as opposed to the vie versa.