Bubble Sort Functionality Update Part 1
-Challenge: showing the numbers as boxes and then swapping them around by comparing them.
Progress: the numbers display as boxes. Struggled at first but that was bc the class i was adding was not in the css script. Figured that out after googling.
Still to complete: Numbers displaying as swapped in the original box design. R.n they do swap but i' trying to figure out how to show the two numbers have been swapped.