Conditional Swap Functionality Update
-Comparing more than two index from the same array.
Key Takeaway:  The condition will always return true if one of the conditions are met , even if the other two were not met. The only condition where the values will not swap if it the comparison between the more than 2 indexes checks if the numbers are equal to each other.