Linear Search Swap Functionality Update
-Figured it out.
The issue : Was using the wrong display method ( the one that was supposed to show the values generally ).
The fix: Created an index that I ascribed to each value in the array and then compared them. In the new display method for the swapped values, I created an array that holds the values of the original array but swapped.